% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_stage.R
\name{md_first_stage}
\alias{md_first_stage}
\title{Run first stage regression
This function is called by the mdqr() function.}
\usage{
md_first_stage(data1, fdep, form1, quantiles, n_small)
}
\arguments{
\item{data1}{data Subsample of dataset containing one group.}

\item{fdep}{formula containing the dependent variable.}

\item{form1}{formula containing endogenous and exogenous regressors.}

\item{quantiles}{vector of quantiles.}

\item{n_small}{Excludes groups with few observations.}
}
\description{
Run first stage regression
This function is called by the mdqr() function.
}
