% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_mdqr}
\alias{plot_mdqr}
\title{Plot results}
\usage{
plot_mdqr(object, x, level = 95)
}
\arguments{
\item{object}{An object of class \code{mdqr}.}

\item{x}{A string containing the name of the regressor of interest}

\item{level}{set confidence level; default is 95 percent.}
}
\value{
A \code{ggplot}
}
\description{
Plot results
}
\section{Additional options}{
Additional \code{\link[ggplot2]{ggplot}} options can be added after the command.
For example: \code{plot_results(md_object, "treatment") +  labs(x = "Quantiles", y = "Point Estimates" , title = "My Title")}
}

