% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(object, x)
}
\arguments{
\item{object}{An object estimated by \code{mdqr}.}

\item{x}{A string containing the name of the regressor of interest}
}
\value{
A \code{ggplot}
}
\description{
Plot results
}
\section{Additional options}{
Additional \code{\link[ggplot2]{ggplot}} options can be added after the command.
For example: \code{plot_results(md_object, "treatment") +  labs(x = "Quantiles", y = "Point Estimates" , title = "My Title")}
}

