% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mdqr.R
\name{print.mdqr}
\alias{print.mdqr}
\title{Print results of mdqr object}
\usage{
\method{print}{mdqr}(x, ...)
}
\arguments{
\item{x}{An object of class mdqr.}

\item{...}{Additional arguments to be passed to the print method.}
}
\value{
The function prints the first element of the mdqr object and returns the object invisibly.
}
\description{
Print results of mdqr object
}
