% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_table.R
\name{result_table}
\alias{result_table}
\title{Create a result table for the regressor of interest}
\usage{
result_table(object, x)
}
\arguments{
\item{object}{An object estimated by \code{mdqr}.}

\item{x}{A string containing the name of the regressor of interest.}
}
\value{
Table of results.
}
\description{
Create a result table for the regressor of interest
}
