% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_table.R
\name{summary_mdqr}
\alias{summary_mdqr}
\title{Create a result table for the regressor of interest}
\usage{
summary_mdqr(object, x = NULL)
}
\arguments{
\item{object}{An object of class \code{mdqr}.}

\item{x}{A string containing the name of the regressor of interest. If left unspecified the regression results for all regressors are returned.}
}
\value{
Table(s) of results.
}
\description{
Create a result table for the regressor of interest
}
