% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mdqr}
\alias{mdqr}
\title{Run minimum distance regression.}
\usage{
mdqr(
  formula,
  data,
  method = c("within", "be", "reoi", "regmm", "ols", "2sls", "gmm"),
  quantiles = seq(0.1, 0.9, 0.1),
  clustervar = NULL,
  cores = NULL,
  n_small = 1,
  run_second = TRUE,
  fitted_values = NULL
)
}
\arguments{
\item{formula}{An object of class \code{\link[Formula]{Formula}}. The formula consists of five parts. \code{y ~ exo_1 + exo_2 | endo_1 + endo_2 | z_1 + z_2 | fe_1 + fe_2 | group_ID }. Note that all part of the formula have to be specified. Use \code{0} to leave it unspecified, e.g.\code{y ~ exo_1 + exo_2 | 0 | 0 | fe_1 + fe_2 | group_ID }. Only second stage fixed effects should be included in the formula. If you wish to estimate a within model you don't need to include fixed effects (see 'details').}

\item{data}{A data.frame containing the necessary variables to run the model.}

\item{method}{A string}

\item{quantiles}{A vector with the quantiles of interest.}

\item{clustervar}{A string with the name of the cluster variable. If \code{clustervar = NULL} (default), group indicator is used as a cluster variable.}

\item{cores}{Number of cores to use for first stage computation. if \code{core = NULL} the number of cores is set to \code{\link[parallel]{detectCores}-1}}

\item{n_small}{A positive integer indicating the minimum size of groups allowed. Groups stricly smaller than \code{n_small} are dropped from the sample.}

\item{run_second}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether the second stage should be performed.}

\item{fitted_values}{A matrix containing the first stage fitted values. To use only if the function \code{mdqr} has been already run and only the second stage is different. For example, to change the clustering of the errors or to change the set of second stage fixed effects.}
}
\value{
A list of two elements. The first element contains regression results for each quantile. The second element contain the matrix of fitted values from the first stage.
}
\description{
Run minimum distance regression.
}
\section{Details}{
NA
}

